requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
discord.py==1.7.3
Django==4.0.5
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
folium==0.12.1.post1
fonttools==4.33.3
future==0.18.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
japanize-matplotlib==1.1.3
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pefile==2022.5.30
Pillow==9.1.1
pip-review==1.1.1
plotly==5.8.2
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMySQL==1.0.2
pyocr==0.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.0
requests-oauthlib==1.3.1
scipy==1.8.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
yarl==1.7.2
